requirements.txt
numpy
pyautogui
screen-brightness-control
ultralytics
pillow
pycaw
comtypes
torch
torchvision
av
imageio
